complete solution using  class components
