+++ src/functionBased/components/InputTodo.js
@@ -1,42 +1,44 @@
-import React, { useState } from "react"
+import React, { Component } from "react"
 
-const InputTodo = props => {
-  const [inputText, setInputText] = useState({
-    title: "",
-  })
+class InputTodo extends Component {
 
-  const onChange = e => {
-    setInputText({
-      ...inputText,
-      [e.target.name]: e.target.value,
-    })
-  }
-
-  const handleSubmit = e => {
-    e.preventDefault()
-    if (inputText.title.trim()) {
-      props.addTodoProps(inputText.title)
-      setInputText({
+    state = {
         title: "",
-      })
-    } else {
-      alert("Please write item")
+        name2: "",
+      };
+
+      onChange = e => {
+        this.setState({
+          [e.target.name]: e.target.value,
+        });
+      };
+
+      handleSubmit = e =>{
+        e.preventDefault();
+        if(this.state.title.trim()){
+        this.props.addTodoProps(this.state.title);
+        this.setState({
+            title: "",
+        })
+      }else{
+        alert('Please write the item');
+      }
     }
+
+  render() {
+    return (
+      <form onSubmit={this.handleSubmit} className="form-container">
+        <input 
+            name = "title" 
+            type="text" 
+            className="input-text"
+            placeholder="Add Todo..." 
+            value={this.state.title} 
+            onChange={this.onChange}
+        />
+        <button className="input-submit">Submit</button>
+      </form>
+    )
   }
-
-  return (
-    <form onSubmit={handleSubmit} className="form-container">
-      <input
-        type="text"
-        className="input-text"
-        placeholder="Add todo..."
-        value={inputText.title}
-        name="title"
-        onChange={onChange}
-      />
-      <button className="input-submit">Submit</button>
-    </form>
-  )
 }
-
 export default InputTodo
