+++ src/functionBased/components/TodoContainer.js
@@ -1,81 +1,100 @@
-import React, { useState,  useEffect } from "react"
-import Header from "./Header"
-import InputTodo from "./InputTodo"
+import React, { Component } from "react";
 import TodosList from "./TodosList"
-import { v4 as uuidv4 } from "uuid"
+import Header from "./Header";
+import InputTodo from "./InputTodo";
+import { v4 as uuidv4 } from "uuid";
 
-const TodoContainer = () => {
-  const [todos, setTodos] = useState([])
-
-  const handleChange = id => {
-    setTodos(prevState =>
-      prevState.map(todo => {
-        if (todo.id === id) {
-          return {
-            ...todo,
-            completed: !todo.completed,
-          }
-        }
-        return todo
-      })
-    )
-  }
-
-  const delTodo = id => {
-    setTodos([
-      ...todos.filter(todo => {
-        return todo.id !== id
-      }),
-    ])
-  }
-
-  const addTodoItem = title => {
-    const newTodo = {
-      id: uuidv4(),
-      title: title,
-      completed: false,
+class TodoContainer extends React.Component{
+    state = {
+        todos : [],
     }
-    setTodos([...todos, newTodo])
-  }
 
-  const setUpdate = (updatedTitle, id) => {
-    setTodos(
-      todos.map(todo => {
-        if (todo.id === id) {
-          todo.title = updatedTitle
+   
+    handleChange = (id) => {
+        this.setState((prevState) => {
+            return{
+            todos:prevState.todos.map(todo => {
+                if(todo.id === id){
+                    return {
+                        ...todo, 
+                        completed: !todo.completed,
+                    };
+            }
+                return todo;
+            })
+        }})
+        
+      };
+     /*  delTodo = (id) => {
+        console.log("deleted", id);
+      } */
+      delTodo = (id) => {
+        this.setState({
+                todos: [
+                    ...this.state.todos.filter(todo =>{
+                        return todo.id !== id;
+                    })
+                ]}
+        )
+
         }
-        return todo
-      })
-    )
-  }
+        addTodoItem = title => {
+            const newTodo = {
+                id: uuidv4(),
+                title: title,
+                completed: false
+              };
 
-  useEffect(() => {
-    // storing todos items
-    const temp = JSON.stringify(todos)
-    localStorage.setItem("todos", temp)
-  }, [todos])
+              this.setState({
+                todos: [...this.state.todos, newTodo]
+              });
+          };
 
-  function getInitialTodos() {
-    // getting stored items
-    const temp = localStorage.getItem("todos")
-    const savedTodos = JSON.parse(temp)
-    return savedTodos || []
-  }
-  
-  return (
-    <div className="container">
-      <div className="inner">
-        <Header />
-        <InputTodo addTodoProps={addTodoItem} />
-        <TodosList
-          todos={todos}
-          handleChangeProps={handleChange}
-          deleteTodoProps={delTodo}
-          setUpdate={setUpdate}
-        />
-      </div>
-    </div>
-  )
+          setUpdate = (updatedTitle, id) => {
+            this.setState({
+                todos: this.state.todos.map(todo =>{
+                    if(todo.id===id){
+                        todo.title = updatedTitle;
+                    }
+                    return todo
+                })
+            })
+            
+          }
+
+          componentDidUpdate(prevProps, prevState) {
+            if(prevState.todos !== this.state.todos) {
+              const temp = JSON.stringify(this.state.todos)
+              localStorage.setItem("todos", temp)
+            }
+          }
+
+          componentDidMount() {
+            const temp = localStorage.getItem("todos")
+            const loadedTodos = JSON.parse(temp)
+            if (loadedTodos) {
+              this.setState({
+                todos: loadedTodos
+              })
+            }
+          }
+          
+    render(){
+        return (
+            <div className="container">
+            <div className="inner">
+                <Header />
+                <InputTodo addTodoProps = {this.addTodoItem}/>
+                <TodosList 
+                todos={this.state.todos} 
+                handleChangeProps = {this.handleChange}
+                deleteTodoProps = {this.delTodo} 
+                setUpdate={this.setUpdate}
+                />
+            </div>
+        </div>
+        )
+    }
 }
 
-export default TodoContainer
+export default TodoContainer;
