+++ src/functionBased/components/TodoItem.js
@@ -1,67 +1,78 @@
-import React, { useState, useEffect } from "react"
-import styles from "./TodoItem.module.css"
+import React from "react";
+import styles from "./TodoItem.module.css";
 
-const TodoItem = props => {
-  const [editing, setEditing] = useState(false)
+class TodoItem extends React.Component{
+    state = {
+        editing: false,
+      }
 
-  const handleEditing = () => {
-    setEditing(true)
-  }
+    handleEditing = () => {
+        this.setState({
+            editing:true,
+        })
+      }
 
-  const handleUpdatedDone = event => {
-    if (event.key === "Enter") {
-      setEditing(false)
+      handleUpdatedDone = (event) => {
+        if(event.key === "Enter"){
+            this.setState({
+                editing: false,
+            })}
+        console.log(event.key);
+      }
+
+      componentWillUnmount() {
+        console.log("Cleaning up...")
+      }
+
+    render(){
+        const completedStyle = {
+            fontStyle: "italic",
+            color: "#595959",
+            opacity: 0.4,
+            textDecoration: "line-through",
+          }
+
+          const { completed, id, title } = this.props.todo;
+
+            let viewMode = {};
+            let editMode = {};
+
+            if (this.state.editing) {
+                viewMode.display = "none";
+            } else {
+                editMode.display = "none";
+            }
+
+        return <li className={styles.item}> 
+            <div onDoubleClick={this.handleEditing} style={viewMode}>
+                <input 
+                type="checkbox" 
+                className={styles.checkbox}
+                checked = {this.props.todo.completed}
+                onChange = {()=>{
+                    this.props.handleChangeProps(id);
+                }}
+                />
+                <button
+                    onClick={() => this.props.deleteTodoProps(id)}>
+                        Delete
+                </button>
+                <span style={completed ? completedStyle : null}>
+                    {title}
+                </span>
+            </div>
+            <input 
+                type="text" 
+                value={title}
+                style={editMode} 
+                className={styles.textInput} 
+                onChange={e => {
+                    this.props.setUpdate(e.target.value, id)
+                  }}
+                  onKeyDown={this.handleUpdatedDone}
+            />
+        </li>
     }
-  }
-
-  const completedStyle = {
-    fontStyle: "italic",
-    color: "#595959",
-    opacity: 0.4,
-    textDecoration: "line-through",
-  }
-
-  const { completed, id, title } = props.todo
-
-  let viewMode = {}
-  let editMode = {}
-
-  if (editing) {
-    viewMode.display = "none"
-  } else {
-    editMode.display = "none"
-  }
-  
-  useEffect(() => {
-    return () => {
-      console.log("Cleaning up...")
-    }
-  }, [])
-
-  return (
-    <li className={styles.item}>
-      <div onDoubleClick={handleEditing} style={viewMode}>
-        <input
-          type="checkbox"
-          className={styles.checkbox}
-          checked={completed}
-          onChange={() => props.handleChangeProps(id)}
-        />
-        <button onClick={() => props.deleteTodoProps(id)}>Delete</button>
-        <span style={completed ? completedStyle : null}>{title}</span>
-      </div>
-      <input
-        type="text"
-        style={editMode}
-        className={styles.textInput}
-        value={title}
-        onChange={e => {
-          props.setUpdate(e.target.value, id)
-        }}
-        onKeyDown={handleUpdatedDone}
-      />
-    </li>
-  )
 }
 
 export default TodoItem
