+++ src/index.js
@@ -1,14 +1,17 @@
 import React from "react"
 import ReactDOM from "react-dom"
+import { BrowserRouter as Router } from "react-router-dom"
 //component file
 import TodoContainer from "./functionBased/components/TodoContainer"
 
 //stylesheet
-import "./functionBased/App.css"
+import "./classBased/App.css"
 
 ReactDOM.render(
-  <React.StrictMode>
-    <TodoContainer />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+    <React.StrictMode>
+      <Router>
+        <TodoContainer />
+      </Router>
+    </React.StrictMode>,
+    document.getElementById("root")
+  )
